added controller with resource

+++ app/Http/Controllers/Admin/HomeController.php
@@ -25,42 +25,35 @@ class HomeController extends Controller
 
 
 
-    public function store(Request $request)
-    {
-        $validated = $request->validate([
-            'title' => 'required|unique:posts|max:255',
-            'body' => 'required',
-        ]);
-    }
-
-
-
-
-
-
-    public function home()
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
     {
         return view('admin.home.home', [
             'data' => $this->data
         ]);
     }
 
-
-
-
-
-
-    public function add()
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
     {
         return view('admin.home.insert');
     }
 
-
-
-
-
-
-    public function insert(Request $request)
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
     {
         $path = $request->file('img')->store('home');
         Home::insert([
@@ -68,33 +61,31 @@ class HomeController extends Controller
             'subtitle' => $request->subtitle,
             'img' => $path,
         ]);
-        return redirect(route('admin.home.view'));
+        return redirect(route('home.index'));
     }
 
-
-
-
-
-
-
-
-    public function delete($id){
-        $item = Home::findorFail($id);
-        $img = $item->img;
-        $item->delete();
-        Storage::delete($img);
-        return redirect(route('admin.home.view'));
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        //
     }
 
-
-
-
-
-
-
-    public function update($id){
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
         $item = Home::findorFail($id);
         return view('admin.home.update',[
+            'item' => $item,
             'data' => $this->data,
             'id' => $id,
             'title' => $item->title,
@@ -103,20 +94,41 @@ class HomeController extends Controller
         ]);
     }
 
-
-
-
-
-
-    public function updateForm(Request $request){
-        Storage::delete($request->imgHid);
-        $path = $request->file('img')->store('home');
-        $item = Home::findorFail($request->id);
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        $img = $request->imgHid;
+        if(!empty($request->img)){
+            Storage::delete($request->imgHid);
+            $img = $request->file('img')->store('home');
+        }
+        $item = Home::findorFail($id);
         $item->update([
             'title' => $request->title,
             'subtitle' => $request->subtitle,
-            'img' => $path
+            'img' => $img
         ]);
-        return redirect(route('admin.home.view'));
+        return redirect(route('home.index'));
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        $item = Home::findorFail($id);
+        $img = $item->img;
+        $item->delete();
+        Storage::delete($img);
+        return redirect(route('home.index'));
     }
 }
